:recycle: Move submission tab to own componenet
