+++ package.json
@@ -25,6 +25,7 @@
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^0.88.2",
     "stylelint": "^13.0.0",
+    "uuid": "^8.1.0",
     "whatwg-fetch": "^3.0.0"
   },
   "scripts": {
