+++ src/components/IdDetails.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {Loader, Tab, Form, Button, Icon} from 'semantic-ui-react';
+import {Loader, Tab} from 'semantic-ui-react';
 import ReactJson from 'react-json-view';
-import DatePicker from 'react-datepicker';
-import 'react-datepicker/dist/react-datepicker.css';
 import {logErrors} from '../utils/common';
 import ReferenceTable from './tables/ReferenceTable';
 import Timeline from './tables/Timeline';
+import SubmitPatientData from './SubmitPatientData';
 import './IdDetails.css';
 
 class IdDetails extends React.Component {
@@ -15,7 +14,6 @@ class IdDetails extends React.Component {
     this.state = {
       loading: false,
       payload: {},
-      observations: [],
       abortController: new AbortController(),
     };
   }
@@ -67,63 +65,6 @@ class IdDetails extends React.Component {
           logErrors(`Error getting schema for ${resourceId}:`, err),
         );
     });
-  };
-
-  addObservation = () => {
-    const newObservation = {type: 'Select a type...'};
-    const observations = [...this.state.observations].concat(newObservation);
-    this.setState({observations});
-  };
-
-  updateObservation = (index, field, value) => {
-    let observations = [...this.state.observations];
-    if (field === 'type' && value !== observations[index].type) {
-      // remove value from object
-      const {value, ...noVal} = observations[index];
-      observations[index] = {...noVal};
-    }
-    observations[index] = {...observations[index], [field]: value};
-    this.setState({observations});
-  };
-
-  getInputType = (event, index) => {
-    switch (event.type) {
-      case 'input':
-        return (
-          <Form.Input
-            placeholder="Please enter a value"
-            value={this.state.observations[index].value || ''}
-            onChange={e =>
-              this.updateObservation(index, 'value', e.target.value)
-            }
-          />
-        );
-      case 'radio':
-        return (
-          <React.Fragment>
-            {event.labels.map((label, i) => (
-              <Form.Radio
-                key={`${label}-${i}`}
-                label={label}
-                value={label}
-                checked={
-                  this.state.observations[index].value &&
-                  this.state.observations[index].value === label
-                }
-                onChange={() => this.updateObservation(index, 'value', label)}
-              />
-            ))}
-          </React.Fragment>
-        );
-      default:
-        return null;
-    }
-  };
-
-  removeObservation = index => {
-    let observations = [...this.state.observations];
-    observations.splice(index, 1);
-    this.setState({observations});
   };
 
   render() {
@@ -200,56 +141,7 @@ class IdDetails extends React.Component {
               <Tab.Pane>
                 Please add any additional observations for this Patient to
                 augment the current dataset.
-                <Form>
-                  {this.state.observations.map((event, i) => {
-                    const details = observationMapping[event.type];
-                    return (
-                      <Form.Group key={`${event.type}-${i}`}>
-                        <Icon
-                          name="delete"
-                          size="large"
-                          onClick={() => this.removeObservation(i)}
-                        />
-                        <Form.Field inline>
-                          <div className="id-details__submit-date-picker field">
-                            <label>Date of Observation</label>
-                            <DatePicker
-                              selected={this.state.observations[i].date}
-                              onChange={date =>
-                                this.updateObservation(i, 'date', date)
-                              }
-                            />
-                          </div>
-                        </Form.Field>
-                        <Form.Field inline>
-                          <Form.Select
-                            label="Observation Type"
-                            options={formOptions}
-                            placeholder={event.type}
-                            onChange={(e, {value}) =>
-                              this.updateObservation(i, 'type', value)
-                            }
-                          />
-                        </Form.Field>
-                        {!!details ? (
-                          <Form.Field inline>
-                            <label>{details.text}</label>
-                            {this.getInputType(details, i)}
-                          </Form.Field>
-                        ) : null}
-                      </Form.Group>
-                    );
-                  })}
-                  <div className="id-details__submit-buttons">
-                    <Button primary onClick={this.addObservation}>
-                      <Icon name="plus" />
-                      Add Observation
-                    </Button>
-                    <Button primary onClick={this.submit}>
-                      Submit
-                    </Button>
-                  </div>
-                </Form>
+                <SubmitPatientData payload={payload} />
               </Tab.Pane>
             ),
           }
